+++ Cargo.toml
@@ -23,7 +23,7 @@ test = false
 
 [dependencies]
 hal-core = { path = "hal-core" }
-mycelium-alloc = { path = "alloc" }
+mycelium-alloc = { path = "alloc", features = ["buddy", "bump"] }
 mycelium-util = { path = "util" }
 rlibc = "1.0"
